python/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
yt-dlp>=2023.10.0
torch>=2.0.0
torchvision>=0.15.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
requests>=2.31.0
pymongo>=4.5.0
python-dotenv>=1.0.0

# Multi-Modal Detection System Requirements
# Pose detection
mediapipe>=0.10.0

# Fire/Smoke detection (CNN)
tensorflow>=2.13.0
keras>=2.13.0

# Audio processing
pyaudio>=0.2.11
librosa>=0.10.0
soundfile>=0.12.0

# Scientific computing
scipy>=1.11.0